uploading as octet-stream
